# Optimize preparation of options for Rouge